Refactor TripDetails Component
