+++ packages/app/screens/trip/TripDetails/styles/styles.tsx
@@ -1,0 +1,23 @@
+export default function loadStyles(theme) {
+  const { currentTheme } = theme;
+  return {
+    mainContainer: {
+      backgroundColor: currentTheme.colors.background,
+      flexDirection: 'column',
+      gap: 15,
+      padding: [25, 25, 0, 25], // [top, right, bottom, left
+      fontSize: 18,
+      width: '100%',
+    },
+    packsContainer: {
+      backgroundColor: currentTheme.colors.cardIconColor,
+      flexDirection: 'column',
+      minHeight: '100vh',
+      padding: 25,
+      fontSize: 26,
+    },
+    dropdown: {
+      backgroundColor: currentTheme.colors.white,
+    },
+  };
+}
